requirements.txt
# LangChain core packages (latest as of Jan 2026)
langchain>=1.2.3
langchain-anthropic>=1.3.1
langchain-community>=0.3.0
langgraph>=1.0.6

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
